♻️  refactor: Updated tests to match rats.cli refactor
